Added 'if currrent_user' for new user signing up

+++ app/controllers/users_controller.rb
@@ -10,12 +10,12 @@ class UsersController < ApplicationController
       if @user.first_name.blank?
         flash[:success] = "You are logged in as a temporary guest. Please be aware that any work you do while logged in as a 'temporary guest' will not be recorded after you have logged out. But if you decide to Join BSC English Online, all your work from this time will be retained."
       else
-        @user.plans = current_user.plans # Guest user's plans transferred to created user before guest user is destroyed.
-        current_user.destroy # guest user record destroyed
+        @user.plans = current_user.plans if current_user # guest user's plans transferred to new student
+        current_user.destroy if current_user # guest user record destroyed
         flash[:success] = "You now have a 'member account' with BSC English Online. Welcome aboard!"
       end
       session[:user_id] = @user.id
-      redirect_to root_path
+      redirect_to home_path
     else
       flash[:danger] = "You were not able to Sign Up"
       render :new
